Create abstract detector base class
