+++ src/detector/detector.py
@@ -6,10 +6,11 @@ import pickle
 import sys
 import tempfile
 import asyncio
-import math
 import numpy as np
 import requests
 from numpy import median
+from abc import ABC, abstractmethod
+import importlib
 
 sys.path.append(os.getcwd())
 from src.base.clickhouse_kafka_sender import ClickHouseKafkaSender
@@ -34,7 +35,7 @@ CONSUME_TOPIC_PREFIX = config["environment"]["kafka_topics_prefix"]["pipeline"][
     "inspector_to_detector"
 ]
 
-
+PLUGIN_PATH = "src.detector.plugins"
 class WrongChecksum(Exception):  # pragma: no cover
     """
     Exception if Checksum is not equal.
@@ -43,7 +44,20 @@ class WrongChecksum(Exception):  # pragma: no cover
     pass
 
 
-class Detector:
+class DetectorAbstractBase(ABC):    
+    @abstractmethod
+    def __init__(self, detector_config, consume_topic) -> None:
+        pass
+
+    @abstractmethod
+    def get_model_download_url(self):
+        pass
+     
+    @abstractmethod
+    def predict(self, message) -> np.ndarray :
+        pass
+
+class DetectorBase(DetectorAbstractBase):
     """Logs detection with probability score of requests. It runs the provided machine learning model.
     In addition, it returns all individually probabilities of the anomalous batch.
     """
@@ -53,7 +67,6 @@ class Detector:
         self.model = detector_config["model"]
         self.checksum = detector_config["checksum"]
         self.threshold = detector_config["threshold"]
-        self.model_base_url = detector_config["base_url"]
 
         self.consume_topic = consume_topic
         self.suspicious_batch_id = None
@@ -150,6 +163,7 @@ class Detector:
                 f"    ⤷  Contains data field of {len(self.messages)} message(s). Belongs to subnet_id {key}."
             )
 
+
     def _sha256sum(self, file_path: str) -> str:
         """Return a SHA265 sum check to validate the model.
 
@@ -178,12 +192,9 @@ class Detector:
         """
         logger.info(f"Get model: {self.model} with checksum {self.checksum}")
         if not os.path.isfile(self.model_path):
-            response = requests.get(
-                f"{self.model_base_url}/files/?p=%2F{self.model}_{self.checksum}.pickle&dl=1"
-            )
-            logger.info(
-                f"{self.model_base_url}/files/?p=%2F{self.model}_{self.checksum}.pickle&dl=1"
-            )
+            download_url = self.get_model_download_url()
+            logger.info(f"downloading model {self.model} from {download_url} with checksum {self.checksum}")
+            response = requests.get(download_url)
             response.raise_for_status()
 
             with open(self.model_path, "wb") as f:
@@ -205,129 +216,12 @@ class Detector:
 
         return clf
 
-    def clear_data(self):
-        """Clears the data in the internal data structures."""
-        self.messages = []
-        self.begin_timestamp = None
-        self.end_timestamp = None
-        self.warnings = []
-
-    def _get_features(self, query: str):
-        """Transform a dataset with new features using numpy.
-
-        Args:
-            query (str): A string to process.
-
-        Returns:
-            dict: Preprocessed data with computed features.
-        """
-
-        # Splitting by dots to calculate label length and max length
-        label_parts = query.split(".")
-        label_length = len(label_parts)
-        label_max = max(len(part) for part in label_parts)
-        label_average = len(query.strip("."))
-
-        logger.debug("Get letter frequency")
-        alc = "abcdefghijklmnopqrstuvwxyz"
-        freq = np.array(
-            [query.lower().count(i) / len(query) if len(query) > 0 else 0 for i in alc]
-        )
-
-        logger.debug("Get full, alpha, special, and numeric count.")
-
-        def calculate_counts(level: str) -> np.ndarray:
-            if len(level) == 0:
-                return np.array([0, 0, 0, 0])
-
-            full_count = len(level)
-            alpha_count = sum(c.isalpha() for c in level) / full_count
-            numeric_count = sum(c.isdigit() for c in level) / full_count
-            special_count = (
-                sum(not c.isalnum() and not c.isspace() for c in level) / full_count
-            )
-
-            return np.array([full_count, alpha_count, numeric_count, special_count])
-
-        levels = {
-            "fqdn": query,
-            "thirdleveldomain": label_parts[0] if len(label_parts) > 2 else "",
-            "secondleveldomain": label_parts[1] if len(label_parts) > 1 else "",
-        }
-        counts = {
-            level: calculate_counts(level_value)
-            for level, level_value in levels.items()
-        }
-
-        logger.debug("Get frequency standard deviation, median, variance, and mean.")
-        freq_std = np.std(freq)
-        freq_var = np.var(freq)
-        freq_median = np.median(freq)
-        freq_mean = np.mean(freq)
-
-        logger.debug(
-            "Get standard deviation, median, variance, and mean for full, alpha, special, and numeric count."
-        )
-        stats = {}
-        for level, count_array in counts.items():
-            stats[f"{level}_std"] = np.std(count_array)
-            stats[f"{level}_var"] = np.var(count_array)
-            stats[f"{level}_median"] = np.median(count_array)
-            stats[f"{level}_mean"] = np.mean(count_array)
-
-        logger.debug("Start entropy calculation")
-
-        def calculate_entropy(s: str) -> float:
-            if len(s) == 0:
-                return 0
-            probabilities = [float(s.count(c)) / len(s) for c in dict.fromkeys(list(s))]
-            entropy = -sum(p * math.log(p, 2) for p in probabilities)
-            return entropy
-
-        entropy = {level: calculate_entropy(value) for level, value in levels.items()}
-
-        logger.debug("Finished entropy calculation")
-
-        # Final feature aggregation as a NumPy array
-        basic_features = np.array([label_length, label_max, label_average])
-        freq_features = np.array([freq_std, freq_var, freq_median, freq_mean])
-
-        # Flatten counts and stats for each level into arrays
-        level_features = np.hstack([counts[level] for level in levels.keys()])
-        stats_features = np.array(
-            [stats[f"{level}_std"] for level in levels.keys()]
-            + [stats[f"{level}_var"] for level in levels.keys()]
-            + [stats[f"{level}_median"] for level in levels.keys()]
-            + [stats[f"{level}_mean"] for level in levels.keys()]
-        )
-
-        # Entropy features
-        entropy_features = np.array([entropy[level] for level in levels.keys()])
-
-        # Concatenate all features into a single numpy array
-        all_features = np.concatenate(
-            [
-                basic_features,
-                freq,
-                freq_features,
-                level_features,
-                stats_features,
-                entropy_features,
-            ]
-        )
-
-        logger.debug("Finished data transformation")
-
-        return all_features.reshape(1, -1)
-
     def detect(self) -> None:  # pragma: no cover
         """Method to detect malicious requests in the network flows"""
         logger.info("Start detecting malicious requests.")
         for message in self.messages:
             # TODO predict all messages
-            y_pred = self.model.predict_proba(
-                self._get_features(message["domain_name"])
-            )
+            y_pred = self.predict(message)
             logger.info(f"Prediction: {y_pred}")
             if np.argmax(y_pred, axis=1) == 1 and y_pred[0][1] > self.threshold:
                 logger.info("Append malicious request to warning.")
@@ -340,6 +234,14 @@ class Detector:
                     "sha256": self.checksum,
                 }
                 self.warnings.append(warning)
+
+
+    def clear_data(self):
+        """Clears the data in the internal data structures."""
+        self.messages = []
+        self.begin_timestamp = None
+        self.end_timestamp = None
+        self.warnings = []
 
     def send_warning(self) -> None:
         """Dispatch warnings saved to the object's warning list"""
@@ -448,6 +350,8 @@ class Detector:
             )
         )
 
+    
+
     def bootstrap_detector_instance(self):
         while True:
             try:
@@ -474,7 +378,7 @@ class Detector:
         loop = asyncio.get_event_loop()
         await loop.run_in_executor(None, self.bootstrap_detector_instance)
 
-
+    
 async def main():  # pragma: no cover
     """
     Creates the :class:`Detector` instance. Starts a loop that continously fetches data.
@@ -486,13 +390,17 @@ async def main():  # pragma: no cover
         KeyboardInterrupt: Execution interrupted by user. Closes down the :class:`LogCollector` instance.
     """
     tasks = []
-    for detector in DETECTORS:
-        consume_topic = f"{CONSUME_TOPIC_PREFIX}-{detector['name']}"
-        detector = Detector(detector_config=detector, consume_topic=consume_topic)
+    for detector_config in DETECTORS:
+        consume_topic = f"{CONSUME_TOPIC_PREFIX}-{detector_config['name']}"
+        class_name = detector_config["detector_class_name"]
+        module_name = f"{PLUGIN_PATH}.{detector_config['detector_module_name']}"
+        module = importlib.import_module(module_name)
+        DetectorClass = getattr(module, class_name)
+        # if not issubclass(DetectorClass, DetectorBase):
+        #     raise TypeError(f"{class_name} is not a subclass of DetectorBase")
+        detector = DetectorClass(detector_config=detector_config, consume_topic=consume_topic)
         tasks.append(asyncio.create_task(detector.start()))
-
     await asyncio.gather(*tasks)
-
 
 if __name__ == "__main__":  # pragma: no cover
     asyncio.run(main())
